big_thing/google_api_big_thing/requirements.txt
google-cloud-vision
pytube
youtube-search-python